Remove the height and width parameters - they are induced from iframe
The iframe that contains the API has to define a height and width.
We assume that the API viewer only will be used inside iframes,
therefore the container will take all the space in the iframe.

Note that we impose the iframe has to be of minimal size 600px x 600px